Removed unnecessary `LINK_LIBRARIES` directive.
With the change in https://github.com/personalrobotics/chimera/pull/159
it should no longer be necessary to explicitly pass `LINK_LIBRARIES`
as an argument to `add_chimera_binding()`.

This should be merged at some point, once we verify that it
actually works for the `dartpy` build.  If it doesn't, we may need
a followup PR to https://github.com/personalrobotics/chimera/pull/159
to fix the issue.